requirements.txt
feedparser
PyYAML
PyGithub
pytube
python-dateutil
pyzotero
markdownify